python-server-ocr/requirements.txt
grpcio
grpcio-tools
easyocr